refactored inventory code
functionality relating to the inventory now stays within the inventory. it's more object oriented i.e the inventory does not "know" what objects the drone sees because it is just an inventory. the inventory does not handle the instancing of items because the "inventory" does not know how to "drop" items, only the drone.